Moved background color logic to the `P5AsciifierManager` class, exposed through the `p5asciify` object
